% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps.R
\docType{data}
\name{crps}
\alias{crps}
\title{Observations of 6-h rainfall amount with CRPS values of 3 calibrated ensemble forecasts for one lead time across France.}
\format{
A matrix with 112221 rows and 4 variables:
\describe{
 \item{obs_rr6}{observations, in mm/6h}
 \item{crps_forecastX}{CRPS values of the forecaster X, in mm/6h}
 ...
}
}
\source{
{Maxime Taillardat}
}
\usage{
crps
}
\description{
Observations of 6-h rainfall amount with CRPS values of 3 calibrated ensemble forecasts for one lead time across France.
}
\keyword{datasets}
