% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexfore.R
\name{indexfore}
\alias{indexfore}
\title{Function for computing the index for a forecast system vs. climatological forecast. You must provide an indexclim object.}
\usage{
indexfore(score_fore, clim)
}
\arguments{
\item{score_fore}{the time serie of the ensemble forecast's CRPS/MAE. Be careful that score_fore is consistent with "score" in indexclim}

\item{clim}{an indexclim object coming from indexclim}
}
\value{
an indexfore object with the index computed vs. climatological forecast and the statistic omega2
}
\description{
Function for computing the index for a forecast system vs. climatological forecast. You must provide an indexclim object.
}
